+++ tests/sqlite.rs
@@ -8,7 +8,7 @@ use teloxide::dispatching::dialogue::{Serializer, SqliteStorage, Storage};
 #[tokio::test(flavor = "multi_thread")]
 async fn test_sqlite_json() {
     let storage =
-        SqliteStorage::open("./test_db1.sqlite", teloxide::dispatching::dialogue::serializer::JSON)
+        SqliteStorage::open("./test_db1.sqlite", teloxide::dispatching::dialogue::serializer::Json)
             .await
             .unwrap();
     test_sqlite(storage).await;
@@ -28,7 +28,7 @@ async fn test_sqlite_bincode() {
 #[tokio::test(flavor = "multi_thread")]
 async fn test_sqlite_cbor() {
     let storage =
-        SqliteStorage::open("./test_db3.sqlite", teloxide::dispatching::dialogue::serializer::CBOR)
+        SqliteStorage::open("./test_db3.sqlite", teloxide::dispatching::dialogue::serializer::Cbor)
             .await
             .unwrap();
     test_sqlite(storage).await;